refactor: move atoms to folder in popup
