+++ apps/extension/src/entrypoints/popup/atoms/auto-translate.ts
@@ -5,46 +5,6 @@ import { getActiveTabUrl } from '@/utils/utils'
 
 type TranslateConfig = Config['translate']
 
-const EMPTY_TAB_URLS = [
-  'about:blank',
-  'chrome://newtab/',
-  'edge://newtab/',
-  'about:newtab',
-]
-
-const EXTENSION_URLS = [
-  'chrome://extensions/',
-  'chrome-extension://',
-  'edge-extension://',
-  'chrome://newtab/',
-  'edge://newtab/',
-]
-
-async function checkIgnoreTab() {
-  if (typeof window !== 'undefined' && browser.tabs) {
-    const tabs = await browser.tabs.query({
-      active: true,
-      currentWindow: true,
-    })
-    const currentTab = tabs[0]
-    const currentUrl = currentTab?.url || ''
-
-    return (
-      EMPTY_TAB_URLS.some(url => currentUrl.includes(url))
-      || EXTENSION_URLS.some(url => currentUrl.includes(url))
-    )
-  }
-  return false
-}
-
-export const isIgnoreTabAtom = atom<boolean>(false)
-
-// Create a derived atom for initialization
-export const initIsIgnoreTabAtom = atom(null, async (get, set) => {
-  const isIgnore = await checkIgnoreTab()
-  set(isIgnoreTabAtom, isIgnore)
-})
-
 // Sync atom to store the checked state
 export const isCurrentSiteInPatternsAtom = atom<boolean>(false)
 
@@ -107,3 +67,5 @@ export const toggleCurrentSiteAtom = atom(
     set(isCurrentSiteInPatternsAtom, checked)
   },
 )
+
+export const isPageTranslatedAtom = atom<boolean>(false)
