--- apps/extension/src/utils/atoms/translation.ts
@@ -1,3 +1,0 @@
-import { atom } from 'jotai'
-
-export const isPageTranslatedAtom = atom<boolean>(false)
